requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
PyMySQL==1.1.0
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
cryptography